dbhub: Set last modified date of file during initial cloning
The dbhub.io server provides us the last modified date of the database
when cloning it. Because we send the last modified date back to the
server when trying to push a database, it makes sense to initially set
the last modified date of the local file to the date provided by the
server.

See issue https://github.com/sqlitebrowser/dbhub.io/issues/101.