Restore QTextEdit widget for RTL text in the Cell Editor
This restores the Qt text widget that was replaced by a QScintilla one in
45c1e2abfd9dd95c8a744c24406b09dff5502e71

QScintilla does not support right-to-left scripts like Arabic, so QTextEdit
is still needed for these languages. Since the QScintilla editor has other
advantages already explained, it is preserved for general text edition.

RTL texts are detected when data is loaded and the Qt editor is
automatically selected. This is done both when loading data from the cell
and when interactively typing RTL characters in the Cell Editor.

All these changes can be undone if a new QScintilla editor supports RTL
texts.

See comments in issue #1793