Allow using extended regular expressions in the find dialogs
This adds support for searching "this|that" to find either "this" or "that"
or using quantifiers like "{number}". The change applies to the
Find/Replace dialog and the search bar.

See https://en.cppreference.com/w/cpp/regex/ecmascript for details of the
syntax.

This raises the requirement for an external version of QScintilla to v2.11.

See issue #1625