Avoid displaying sort indicators in plot dock and "Copy with headers"
headerData() now returns column name plus sort indicator in the display
role and only the column name in the edit role. This allows to use the
edit role for the plot and the copy-with-headers features, so they do not
show the sort indicator as part of the column name.

See related issue #1409 and PR #1810