grammar: Fix whitespace problems
When using a "x IS NOT NULL" expression in a statement our parser was
generating something like "xIS NOTNULL". This was especially a problem
because the table looked like it parsed correctly but actually contained
a faulty expression. So when modifying the table you would get
unexpected error messages, or worse silent errors introduced into your
table.

Also add a test case for this and for commit e7ba79f47878fe01c9fa35d1c51d59c10669b794.

See issue #1969.