Add lines representing relations to diagram
When a new table is added to the diagram, its relations to parent and
child tables are added too. The lines connecting the tables have start and
end positions based on the table proxies representing the table in the
diagram.

When the tables are moved, their relation lines are moved too.