Fix retrieving list of selected columns for Browse Data tab
When retrieving the list of selected columns of a table in the Browse
Data tab only take into account columns which are selected entirely, not
columns which have at least a single selected field. Before selecting a
row would implicity select all columns.

This fixes issues with resizing and with hiding a column.

See issue #1999.