Remove close buttons on tabs and add button for closing SQL tabs
This gives us some more space in the tab bar and should reduce the risk
of accidentally closing a tab. It looks prettier too. We could add an
option for this in the Preferences dialog but in order to not add too
many rarely used options, let's try it without an option first.

See issue #1968.