Open URL or filename: relative filenames are based on DB path and refactor
When opening a relative filename from "Edit DB Cell" or Table Browser the
relative paths are interpreted from the DB file location. This allows
portability of DB and linked files to a new location.

Refactoring so the open URL code is centralised in MainWindow and a status
message is displayed with error or opening notification.

See issue #1597