Enhancement: save to temporary file, open external application and reload
A new button has been added in "Edit DB Cell" dock, which saves the cell
data to a temporary file with extension according to detected data and
opens default external application for the file type. It then asks user to
reload the data in a dialog when they have finished editing the data. It
can be also used for viewing if the user cancels the reload.

See related issues #1791 and #1746