Take into account Windows line separator in Execute SQL
QScintilla text(line) returns the string of the passed line including the
line terminator. Given the differences in platforms in that regard, we
only supported the case for a single character separator.

This should solve issue #2073 for Windows.

It also should solve #1768 and #1632, which were only reproduced in Windows
or editing files with lines ending in "\r\n" in general.