Fix the recordset numbers in the Browse Data tab
This improves the accuracy of the number of the last row in the
recordset number label at the bottom of the Browse Data tab by using the
seemingly more accurate visualIndexAt approach instead of the rowAt
approach from before and by taking the horizontal scrollbar into
account.

It also updates the label when resizing the row heights or column
widths.

See issue #232.