Disable editing of generated columns in Browse Data tab
This disables all cells in generated columns for editing in the Browse
Data tab. Since the values of these columns are generated by SQLite
there is no way to edit them directly and each attempt will result in an
error. So it is best to not allow editing them in the UI.

See issue #2107.