Fall back to asking SQLite when parsing a table schema has failed
When parsing a table schema has failed, fall back to the PRAGMA provided
by SQLite to give us some information about the layout of the table.
This does not give us as much information but it is definitely better
than no information at all. The main aim here is to fix the case where
we are dealing with a virtual table which we failed to parse, and we
now do not even know it is a virtual table.

See issue #2187.