Avoid unchanged translation of keyboard shortcuts to work around problems
There are some problems with some translations where the keyboard shortcut
does not work when translated. To prevent those problems, a general
reset of all the translations which are supposed to not change the actual
key combination has been made.

Only some shortcuts in Italian are clearly changed by the translation and
those have been preserved.

See issue #2178
as is.