DbStructureQItemViewFacade implemented
DbStructureQItemViewFacade provides a simple read only interface to a
QAbstractItemView (e.g. QTreeView) that holds a model() of type
DBStructureModel.

It is designed to simplify the access to the actual selected node.

The Class follows the facade design pattern.
But it doesn't control the lifecycle of the itemView it is connected to.