Remove "NULL" in margin of editor in Edit dialog for empty cells
When clicking on a cell with a NULL value while, the Edit dialog
switches into text mode and shows an empty text with "NULL" written in
the line number margin. When then changing to an empty string cell, the
text in the editor did not change (it is still empty after all). This
prevented an action from triggering which meant that the "NULL" text in
the line number marging was not removed.

See issue #2204.