Do not show fields twice when parsing a table was incomplete
When the CREATE TABLE statement of a table could not be parsed
completely we asked SQLite for more information on this table and added
that to the table representation. This however made all fields appear
twice. With this commit we only ask SQLite for information on the table
when our parser did not yield any information at all to avoid this
problem.

See issue #2269.