Fix empty filters in project files
Fix loading of project files with empty filter values. If an empty
filter value was saved in a project file, it did not show up in the UI
but was partially added to the query anyway. This lead to an invalid
query with no results.

Also fix a possible loophole which at least in theory could lead to an
empty filter value being stored in the project file.

See issue #2288.