parser: Extended support for NOT NULL and UNIQUE column constraints
This adds support for conflict actions in NOT NULL and UNIQUE column
constraints. It also supports named NOT NULL and UNIQUE column
constraints. With this it is now possible to edit tables which make use
of these constraints without losing any information. Adding constraints
like this via the UI is still not supported though.

Once more constraints are handled like this, the code can be probably
simplified a lot.