Add new option to freeze columns in the Browse Data tab
This adds a new context menu action called "Freeze Columns" to the
context menu which appears when you right click the column headers in
a Browse Data dock. With this action all columns from the first up to
the clicked column are fixed in place when you scroll horizontally. This
can be used to make for example the id column always visible.

See issue #1888.