dbhub: Fix port number mismatch when using port 443
This fixes checking out databases and pushing changes to them later.
Because for the first the port number was omitted and for the second it
was expected, looking up the last commit id in the database returned no
results which led to an 'Update required' error from the dbhub server.