dbhub: Only provide commit id when pushing when the database is a clone
When pushing a database to the remote server only provide the commit id
when the database file is saved in the directory of cloned databases.
This makes sure we do not provide any random commit id when a database
with the same name as a cloned database is pushed.

See issue #1164.