dbhub: Do not allow pushing in-memory databases
Because there is no local file on the disk to upload in-memory database
cannot be uploaded to the remote server. This commit makes sure the
button to upload databases is disabled for in-memory databases. It also
fixes a problem where some information on a previously opened database
would remain visible when opening an in-memory database.