Better error messages when CSV import fails due to an unexpected EOF
When the CSV import fails because of an unexpected end of file show a
better error message. When the file ends but we are still expected more
data to come, i.e. we are still in quote mode, no INSERT statement
failed. So showing the last SQLite error message will never show you a
proper error message. Instead tell the user that their settings might be
wrong or the file malformed.

See issue #2330.