dbhub: Fix sporadic crashes when selecting an identity
This fixes a sporadic crash when selecting an identity in the Remote
dock. It happened because selecting an identity removes the first dummy
entry in the drop down list of identities which for some reason emits
the signal which then calls the same function to update the root dir
again. Depending on the timing of the network requests this could lead
to some confusion when building the tree structure.

Also fix an off-by-one error when building the tree structure.

See issue #2333.