dbhub: Add context menu action for saving a revision to another file
This adds a new context menu action to the commit list in the Current
Database tab of the Remote dock which downloads the database from that
commit and saves it to another file. This file can then be opened and
inspected just like any other database file and does not update the
local clones data.