New predefined settings for Window Layout
A new predefined simplified layout has been added with the only the basic
functions: schema, browsing and SQL execution. Ideal for courses.

Also added options for moving all the docked windows to the bottom, left
side or top (for the right, the default can be restored with Reset Window
Layout as before).

See issue #2307