Add new action to close project file and return to initial state
This will close all SQL tabs, asking if they have to be saved and close
the database file.

Fixed a bug when switching between two project files from the Recent Files
list. If the first project had a pending change and it was saved at that
very moment, the database was already closed and the project file was saved
without DB file.

See issue #2289