Allow multiple TableBrowser tabs in the Browse Data tab
This adds a new tab widget inside the Browse Data tab which allows you
to create multiple TableBrowser widgets. This way you can browse
multiple tables at the same time and switch between them.

I don't consider this done. The UI for this is likely to change in the
future. This commit is just a first step towards this goal.

See issue #2283.