Avoid possible database busy message when using the Browse Table action
Avoid a possible database busy warning when using the Browse Table
context menu action in the Database Structure tab. This happened because
of an unnecessary schema update.

Also do not refresh the view when changing between the main tabs. This
makes tab switching faster, makes sure the scroll position is not lost,
and avoid two queries instead of one when using the Browse Table action.

See issue #2283.