Add more options for setting font sizes
This adds two more options to the Preferences dialog which allow you to
change the font sizes. One option allows changing the font size of the
tree widgets in the Database Stucture tab and the Database Schema dock.
The other option allows changing the font size of all UI elements for
which there is no other font size option.

See issue #2306.

# Conflicts:
#	src/PreferencesDialog.ui