Modify code according to review
Implement setter(setUserConfigurationFile) to store user configuration file.
Move the code that print error when the argument for the option is not provided.
Print detailed information when the value of an environment variable is substituted.
Replaced with 'nullptr' instead of 'NULL' according to Modern C++ standard.
Add code to check if provided file is a normal configuration file.
Change the QSetting class object to be created only once(setSettingsObject).
Using the '->' operator.

Thank you so much for reviewing! @mgrojo