Added message box for informative command line arguments for Windows
Added a message handler so any Qt message sent before the main window is
visible will be shown to user in a message box. In principle, this is
designed for the informative command line arguments: '-h' and '-v', but
will improve also the error messages about not found arguments or any other
warning or error shown before the messages can be seen in the Log dock.

See issue #1069

# Conflicts:
#	src/main.cpp