Fixed incorrect display of SQL calltips containing non-US-ASCII characters
There is a further problem when the calltip is displayed after clicking
one of the arrows for overloaded functions.

This was first reported in #1107 (Russian) and #1206 (Korean). The subcase
has been reported in PR #2424.

Patch reported in the QScintilla list, so there is no need to keep a new
patch. It is assumed to come in new QScintilla versions.