Allow selecting fixed format for big integer numbers
The default format changes to scientific format when the width of the
number would be longer. The user will be able to select fixed format for
integer numbers in the contextual menu. Note that precision is set to 0,
because in the fixed format trailing zeros are not omitted.

See issue #2498