Enable URI filenames in SQLite
Enable usage of URI filenames in SQLite calls. In theory, this should
not break any current use case as we tend to pass full paths to SQLite
and they should never start with "file:" or similar.

This should make it possible to attach databases using full URI
specifiers.

See issue #2493.