CSV Import: give option to use system locale conventions
This will allow to import numbers in CSV which use ',' as decimal separator
and '.' as thousands separator, when the system locale is Spanish (Spain),
for example.

Additionally this commit detects the full range supported by SQLite for
INTEGER and REAL numbers, that is, 64 bit integers are not converted to
REAL and double precision reals are not converted to TEXT.

See issues #1347 and #2140.