Avoid running SELECT statements twice in Execute SQL tab
This speeds up executing SELECT statements in the Execute SQL tab even
further. Before this the statement was executed twice: once to determine
whether it returns any data and another time to actually get the data.
This is still done after this commit, however the process to determine
whether a statement returns any data is optimised in a way which almost
reduces the runtime of this step to zero.

See issue #2165.