Import: Add import-csv option to command line to import CSV files
The argument can be passed several times and all the CSV files are added to
the Import CSV dialog. When no database to open is passed in the command line,
the CSV files are imported into a new in-memory database, which could later
be saved as a file, if desired.

This option could be used as basis for adding a file association to CSV files
for DB4S.

See issue #2589.