Avoid error messages when checking for updates
When using rare network settings errors might occur when checking for
updates. Instead of displaying an irritating error message when starting
the application, just silently ignore errors. This is far from perfect
but at least does not make things worse because even if you see the
error message this commit hides, you would not be able to do much about
it.

See issue #2759.