Fix changes to constraints in Edit Table dialog not always being applied
Changes to constraints in the Edit Table dialog were not always
correctly applied, especially when only constraints were changed. This
was because we did not make deep copies of the constraint objects so
instead of copied we were changing the original constraints. This meant
that the ALTER TABLE code we have in place thought that there were no
changes made in compariron to the original table.