build: Use cmake for macOS builds
This is a first attempt to use cmake instead of qmake for building our
nightlies on macOS. We are using cmake on most platforms and it would be
easier for us to only maintain one build system. Because support for
qmake is dropped in Qt6 anyway it makes sense to transition to only
using cmake in the near-ish future.

Please note that this commit has been assembled in a mostly blind
fashion. I would be surprised if this worked without further tweaking.