SqliteTableModel: Add maximum rows limit
This was implemented as a workaround for #2592 and #2787.
There is an issue with QTableView integer overflow (int32_t), when
QTableView::defaultRowSize() = 30, maximum number of rows which can
be rendered is 0x7fffffff / 30 = 71,582,788. When a huge SQLite table is
selected in the BrowserTab, DB4S hangs in an infinite render loop.

More info https://github.com/sqlitebrowser/sqlitebrowser/issues/2592#issuecomment-896794988