Simplify dependencies and remove unneeded dependencies from tests
This moves the code to remove comments from SQL statements from the
SqliteTableModel class to Data.cpp making it a free function. This
removes some dependencies from the SqliteTableModel class with all its
dependencies.

Also simplify the CMakeLists.txt file for the tests by removing all the
dependencies which are not really required.