Fix browsing tables without rowid and multiple primary key columns
This fixes a possible crash when browsing a table without a rowid column
and with multiple columns in its primary key. To work with these tables
we need to pretend it still has only one primary key column (as in the
rowid case) and to do so we combine all the values of the primary key
columns into a single value. For simplicity we were using the JSON
library (which we are using anyway in the project) to achieve this,
however this turned out to be problematic for some values. This commit
introduced a new, more robust approach to combine multiple values into a
single value.

See issue #2832.